Backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pydantic>=2.7.0
pydantic-settings>=2.3.0
email-validator>=2.1.1
redis>=5.0.0
kafka-python>=2.0.2
